+++ campaign/views.py
@@ -5,7 +5,7 @@ from .models import FbCampaign
 from .serializers import FbCampaignWithAdsBrief
 from ad.models import Ad
 from ad.serializers import AdSerializer
-from .mixins import OrderingMixin
+from utils.mixins import OrderingMixin
 from utils.pagination import CustomPagination
